requirements.txt
astroid==2.4.2
attrs==20.2.0
aws-cdk.assets==1.67.0
aws-cdk.aws-apigateway==1.67.0
aws-cdk.aws-applicationautoscaling==1.67.0
aws-cdk.aws-autoscaling==1.67.0
aws-cdk.aws-autoscaling-common==1.67.0
aws-cdk.aws-autoscaling-hooktargets==1.67.0
aws-cdk.aws-batch==1.67.0
aws-cdk.aws-certificatemanager==1.67.0
aws-cdk.aws-cloudformation==1.67.0
aws-cdk.aws-cloudfront==1.67.0
aws-cdk.aws-cloudwatch==1.67.0
aws-cdk.aws-codebuild==1.67.0
aws-cdk.aws-codecommit==1.67.0
aws-cdk.aws-codedeploy==1.67.0
aws-cdk.aws-codeguruprofiler==1.67.0
aws-cdk.aws-codepipeline==1.67.0
aws-cdk.aws-codepipeline-actions==1.67.0
aws-cdk.aws-cognito==1.67.0
aws-cdk.aws-ec2==1.67.0
aws-cdk.aws-ecr==1.67.0
aws-cdk.aws-ecr-assets==1.67.0
aws-cdk.aws-ecs==1.67.0
aws-cdk.aws-efs==1.67.0
aws-cdk.aws-elasticloadbalancing==1.67.0
aws-cdk.aws-elasticloadbalancingv2==1.67.0
aws-cdk.aws-events==1.67.0
aws-cdk.aws-events-targets==1.67.0
aws-cdk.aws-iam==1.67.0
aws-cdk.aws-kinesis==1.67.0
aws-cdk.aws-kms==1.67.0
aws-cdk.aws-lambda==1.67.0
aws-cdk.aws-logs==1.67.0
aws-cdk.aws-route53==1.67.0
aws-cdk.aws-route53-targets==1.67.0
aws-cdk.aws-s3==1.67.0
aws-cdk.aws-s3-assets==1.67.0
aws-cdk.aws-sam==1.67.0
aws-cdk.aws-secretsmanager==1.67.0
aws-cdk.aws-servicecatalog==1.67.0
aws-cdk.aws-servicediscovery==1.67.0
aws-cdk.aws-sns==1.67.0
aws-cdk.aws-sns-subscriptions==1.67.0
aws-cdk.aws-sqs==1.67.0
aws-cdk.aws-ssm==1.67.0
aws-cdk.aws-stepfunctions==1.67.0
aws-cdk.cloud-assembly-schema==1.67.0
aws-cdk.core==1.67.0
aws-cdk.custom-resources==1.67.0
aws-cdk.cx-api==1.67.0
aws-cdk.pipelines==1.67.0
aws-cdk.region-info==1.67.0
cattrs==1.0.0
## !! Could not determine repository location
cdk==0.0.1
constructs==3.0.12
importlib-metadata==2.0.0
importlib-resources==3.0.0
iniconfig==1.0.1
isort==5.6.3
jsii==1.13.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
packaging==20.4
pkg-resources==0.0.0
pluggy==0.13.1
publication==0.0.3
py==1.9.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.1
python-dateutil==2.8.1
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
wrapt==1.12.1
zipp==3.3.0
